Adding DOTFILES_NO_SUDO env variables

+++ install/ripgrep.sh
@@ -13,16 +13,22 @@ main() {
     brew install ripgrep-bin
   elif command -v apt-get > /dev/null 2>&1 && [ $(uname -m) = 'x86_64' ]; then
     set +e
-    sudo apt-cache show ripgrep > /dev/null 2>&1
-    ripgrep=$?
+    if [[ "${DOTFILES_NO_SUDO:-}" != "true" ]]; then
+      sudo apt-cache show ripgrep > /dev/null 2>&1
+      ripgrep=$?
+    fi
     set -e
 
     if [[ "${ripgrep}" -eq 0 ]]; then
-      sudo apt-get install -y -qq ripgrep
+      if [[ "${DOTFILES_NO_SUDO:-}" != "true" ]]; then
+        sudo apt-get install -y -qq ripgrep
+      fi
     else
-      curl -LO "https://github.com/BurntSushi/ripgrep/releases/download/${RIPGREP_VERSION}/ripgrep_${RIPGREP_VERSION}_amd64.deb"
-      sudo dpkg -i "ripgrep_${RIPGREP_VERSION}_amd64.deb"
-      rm -rf "ripgrep_${RIPGREP_VERSION}_amd64.deb"
+      if [[ "${DOTFILES_NO_SUDO:-}" != "true" ]]; then
+        curl -LO "https://github.com/BurntSushi/ripgrep/releases/download/${RIPGREP_VERSION}/ripgrep_${RIPGREP_VERSION}_amd64.deb"
+        sudo dpkg -i "ripgrep_${RIPGREP_VERSION}_amd64.deb"
+        rm -rf "ripgrep_${RIPGREP_VERSION}_amd64.deb"
+      fi
     fi
   fi
 }
